Remove rgb.h, refactor config references
Move most of the functionality in rgb.h to lamp.h, which better reflects
what it actually does. As part of this, create device/specs.h as the One
True Way to reference configuration. This includes the existing config.h
file and also defines properties that aren't actually configurable, like
the lamp count.