Add OUT interrupt endpoint to USB descriptor
The majority of the device function happens via output reports and since
these might be part of animations, it seems best to send them over a
dedicated lower-latency pipe.

Also clean up some remaining TODOs in the USB code, verifying that all
the flags passed to things are correct, or at least documented.