Send keepalive reports to block USB suspend
Windows 11 seems to selectively suspend USB HID devices that are idle
for 5-10s. We want to get a suspend event only when the system goes to
sleep, so send perioding messages to make Windows think we're doing
something.

Theoretically you can disable this behavior in the OS, but I couldn't
get it to work and having to specially configure something that is
otherwise plug-n-play in Device Manager is annoying.