Add temperature sensor to HID report
The primary purpose for this usage is to trick Windows into leaving the
device in the active state until there is a global suspend as part of
going to sleep. By default, Windows selectively suspends the device when
the bus is idle for a few seconds. I believe this is because nothing
uses the lighting or vendor pages by default, so there are no open
connections to the device. To avoid a special driver or service that's
always running, declare a usage that Windows opens by default [1].

Temperature is a good choice because the RP2040 has an internal sensor,
so I can actually provide data with low effort instead of lying about
usages and having an approximate measure of case ambient temperature
could be handy.

Microsoft also provides a useful whitepaper explaining how to implement
the sensors usage page [2]. Most of the code here is based on that.

Finally, re-number all the reports so I can add new reports without
changing the IDs of everything that happened to be defined later.

[1] https://learn.microsoft.com/en-us/windows-hardware/drivers/hid
      /top-level-collections-opened-by-windows-for-system-use

[2] https://learn.microsoft.com/en-us/windows-hardware/design
      /whitepapers/hid-sensors-usages