Clamp RGB values when converting from oklab
I neglected the fact that when transitioning between certain colors in
oklab space, the resulting RGB values may be outside of [0.0, 1.0]. This
meant that the naive cast to uint8_t overflowed and created weird jumps
in color. Cast to a uint16_t instead, then clamp to the uint8_t range.