Fix resume behavior
For some reason, possibly because of the sensor usage or maybe due to a
tinyUSB bug, we don't get the resume callback when Windows wakes from
sleep. Instead, it looks like we get a bus reset, which triggers the
mount callback. Treat this as equivalent to the resume event. Because
the device is bus-powered and internal, we don't really care about mount
and unmount otherwise.

Now that suspending and resuming works, block on events while suspended
to minimize power usage.