Update lighting reports to better match spec
I found an example descriptor in the original proposal for this page
that indicates many of these reports should be feature reports, not
input/output reports and also pointed out some places where I got the
field sizes wrong.

I'm not sure if these are actually critical, but Windows is refusing to
open my device, and one theory is that something is wrong with my
descriptors. At least in the sensor usage, Windows wouldn't recognize
the device until the descriptor had certain fields.