Convert oklab values directly to 16-bit linear RGB
Oklab is already a perceptual color space, which means when converted to
linear RGB, there's no need to apply a gamma transform to get a
perceptually uniform output. As a result, we can also take advantage of
the full 16-bit PWM range instead of being limited to 8 bits.

8-bit values from Lighting & Illumination HID reports are still squared
to approximate the gamma transform. These are specified to cover a
perceptually linear range, so the nonlinear transfor is required.