Fix issue with off state in breathe animations
I misunderstood how the Oklab space works. While L=0 does mean "zero
lightness", if a and b are non-zero, the resulting color is outside of
the normal RGB space and may not be physically possible. I don't think
it is possible to fade to true black while going "through" some other
color without adding another stop in the gradient. I don't want to add
support for that right now, so simplify the animation to fade between
the two literal colors, where the "off" color is usually black.