Fix lighting HID descriptor for Windows (#3)
The Windows 11 implementation of the Lighting & Illumination page
apparently requires that devices report a descriptor that exactly
matches the example given in the original HUTRR84 proposal. When given a
descriptor that does not match this example, the Windows APIs just hang,
which is not very helpful for debugging.

While I think my implementation was largely correct given what the HID
specs say, here's what I had to change:

* Always use feature reports, do not mix in output reports
* Use 16 bits for LampID usages
* Use 8 bits for LampUpdateFlags usages
* Fix the maximum value of the AutonomousMode usage
* Mark the lamp ID usage of the LampAttributesRequestReport as data
* Redefine the minium value with each usage

I tried a few combinations of these fixes and saw the hang behavior
every time, so my conclusion for now is that they are all required. The
only change that seems acceptable is the number of lamps that can be
updated in a LampMultiUpdateReport.