Temporarily disable lamp-array CLI functions
Now that the device correctly reports as a HID LampArray, Windows hides
it from the generic HID interface. Since the LampArray interface
operates at a higher-level, it's not a direct replacement.

I think I'll have to drop down to the Win32 HID/USB APIs if I want this
to work. For now, just disable these commands since trying to load the
HID device returns an error and breaks everything else.

Most of this functionality is accessible through the Dynamic Lighting
settings page anyway.